requirements.txt
numpy
scipy
scikit-learn
python-igraph
networkx
pyper
-e git+https://github.com/fabriziocosta/EDeN.git#egg=eden
